--- src/game-servers/dto/game-server.dto.ts
@@ -1,7 +1,0 @@
-export interface GameServerDto {
-  id: string;
-  createdAt: Date;
-  name: string;
-  address: string;
-  port: string;
-}
