+++ src/game-servers/game-server-provider.ts
@@ -1,12 +1,8 @@
-import { GameServer } from './models/game-server';
-
-type GameServerConstructor = {
-  new (...args: any[]): GameServer;
-};
+import { GameServerOption } from './interfaces/game-server-option';
 
 export interface GameServerProvider {
   readonly gameServerProviderName: string;
-  readonly implementingClass: GameServerConstructor;
   readonly priority?: number;
-  findFirstFreeGameServer: () => Promise<GameServer>;
+  findAllOptions: () => Promise<GameServerOption[]>;
+  takeServer: (option: GameServerOption) => Promise<void>;
 }
