+++ src/game-servers/interfaces/game-server-option.ts
@@ -1,0 +1,4 @@
+export interface GameServerOption {
+  id: string;
+  name: string;
+}
