+++ src/game-servers/providers/static-game-server/models/static-game-server.ts
@@ -1,23 +1,47 @@
-import { GameServer } from '@/game-servers/models/game-server';
 import { svLogsecret } from '@/game-coordinator/utils/rcon-commands';
 import { createRcon } from '@/utils/create-rcon';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Exclude } from 'class-transformer';
-import { Document } from 'mongoose';
+import { Transform } from 'class-transformer';
+import { Document, Types } from 'mongoose';
 import { Rcon } from 'rcon-client';
-import { staticGameServerProviderName } from '../static-game-server-provider-name';
 import { generateLogsecret } from '../utils/generate-logsecret';
+import { TransformObjectId } from '@/shared/decorators/transform-object-id';
 
 @Schema()
-export class StaticGameServer extends GameServer {
+export class StaticGameServer {
+  __v?: number;
+
+  @TransformObjectId()
+  _id?: Types.ObjectId;
+
+  @Transform(({ value, obj }) => value ?? obj._id?.toString())
+  id!: string;
+
+  @Prop({ default: () => new Date() })
+  createdAt!: Date;
+
+  @Prop({ required: true, trim: true })
+  name: string;
+
+  /**
+   * The gameserver's public IP address.
+   */
+  @Prop({ required: true, trim: true })
+  address: string;
+
+  @Prop({ required: true })
+  port: string;
+
+  @TransformObjectId()
+  @Prop({ type: Types.ObjectId, ref: 'Game' })
+  game?: Types.ObjectId; // currently running game
+
   /**
    * The IP address of the gameserver's that the heartbeat came from.
    */
-  @Exclude({ toPlainOnly: true })
   @Prop()
   internalIpAddress: string;
 
-  @Exclude({ toPlainOnly: true })
   @Prop({ required: true })
   rconPassword!: string;
 
@@ -27,23 +51,11 @@ export class StaticGameServer extends GameServer {
   @Prop({ default: false })
   isOnline!: boolean;
 
-  /**
-   * Was the server cleaned up after the last game?
-   */
-  @Prop({ default: true })
-  isClean!: boolean;
-
   @Prop({ default: () => new Date() })
   lastHeartbeatAt?: Date;
 
   @Prop({ default: 1 })
   priority!: number;
-
-  /**
-   * Set by the sm_tf2pickuporg_voice_channel_name cvar.
-   */
-  @Prop()
-  customVoiceChannelName?: string;
 
   async rcon(): Promise<Rcon> {
     return await createRcon({
@@ -69,9 +81,3 @@ export class StaticGameServer extends GameServer {
 export type StaticGameServerDocument = StaticGameServer & Document;
 export const staticGameServerSchema =
   SchemaFactory.createForClass(StaticGameServer);
-
-export function isStaticGameServer(
-  gameServer: GameServer,
-): gameServer is StaticGameServer {
-  return gameServer.provider === staticGameServerProviderName;
-}
